# lib/api-zod/src/generated/types/createCustomFolderInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateCustomFolderInput {
  /** Client-generated 'cf_<base36>_<rand>' */
  id: string;
  name: string;
}