# lib/api-zod/src/generated/types/customFolderRecord.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

/**
 * A ClinAdmin custom folder definition. No email content.
 */
export interface CustomFolderRecord {
  id: string;
  name: string;
  createdAt: Date;
}