# lib/api-zod/src/generated/types/emailFolderAssignmentRecord.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

/**
 * Reference-only mapping; never carries email content.
 */
export interface EmailFolderAssignmentRecord {
  outlookEmailId: string;
  customFolderId: string;
  assignedAt: Date;
}