# lib/api-zod/src/generated/types/outlookMessageRecord.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

/**
 * Live read through the email-fetch adapter. Subject, sender, snippet, received date only — no body persisted by our app.
 */
export interface OutlookMessageRecord {
  id: string;
  from: string;
  subject: string;
  snippet: string;
  receivedAt: Date;
}